refactor: rename components/effects to components/providers
WebMCPProvider は effect ではなく provider であり、アニメーション層の
撤去後は effects/ に1ファイルだけ取り残される形になっていた。
Frontend App パターンに合わせて providers/ に移動する。
